Fixed again a bug in calculation of the uniform distribution chi square test (Mamma Mia!) Added a handy general-purpose 32-bit hash code generator based on the fast and efficient Fowler-Noll-Vo algorithm. Added more tests.

+++ src/MbUnit/MbUnit/Framework/ContractVerifiers/HashCodeAcceptanceContract.cs
@@ -175,7 +175,7 @@ namespace MbUnit.Framework.ContractVerifiers
         {
             return new TestCase("CollisionProbabilityTest", () => AssertionHelper.Verify(() =>
             {
-                var probability = (double)store.CollisionProbability;
+                var probability = store.CollisionProbability;
                 TestLog.WriteLine("Statistical Population = {0}", store.StatisticalPopulation);
                 TestLog.WriteLine("Actual Collision Probability = {0} {1} {2}", probability, ((probability <= collisionProbabilityLimit) ? "≤" : ">"), collisionProbabilityLimit);
